refactor: place모델 파일명 변경, Button props 수정
